requirements.txt
pyzmq
nose
fabric
pymongo
psutil
